use ocaml-ci-scripts for build; test for 4.04/4.05/4.06
For CI, use ocaml-ci-scripts's .travis-docker.sh and the
.travis.yml-based environment variables to build the environment.

The OCAML_VERSIONs specified are ones with AFL instrumentation enabled by
default, and the images that .travis-docker.sh maps them to have
afl-fuzz, afl-gotcpu, and afl-whatsup installed.  These custom images
are provided by the `hub_user` set in .travis.yml; `hub_user` itself is
a custom bit of configuration only available because .travis-docker.sh
comes from a custom fork.

`hub_user` will be submitted to ocaml-ci-scripts in the short term, allowing the
maintainer of this repository not to track ocaml-ci-scripts on her fork
if the change is merged.
Base images with afl instrumentation will hopefully be available from
ocaml/opam in the future.  It is hoped that afl itself will be available
through opam soon.

Crowbar and Bun will someday be released, allowing the maintainer to
remove pins here.